+++ sdk/search/azure-search-documents/azure/search/documents/indexes/_internal/aio/_search_index_client.py
@@ -9,13 +9,18 @@ from azure.core import MatchConditions
 from azure.core.tracing.decorator_async import distributed_trace_async
 from azure.core.async_paging import AsyncItemPaged
 from .._generated.aio import SearchServiceClient as _SearchServiceClient
+from .._generated.models import SynonymMap
+from ....aio import SearchClient
 from .._utils import (
     delistize_flags_for_index,
     listize_flags_for_index,
+    listize_synonyms,
+    pack_search_resource_encryption_key,
     get_access_conditions,
+    normalize_endpoint,
 )
-from ..._headers_mixin import HeadersMixin
-from ..._version import SDK_MONIKER
+from ...._headers_mixin import HeadersMixin
+from ...._version import SDK_MONIKER
 
 if TYPE_CHECKING:
     # pylint:disable=unused-import,ungrouped-imports
@@ -24,7 +29,7 @@ if TYPE_CHECKING:
     from azure.core.credentials import AzureKeyCredential
 
 
-class SearchIndexesClient(HeadersMixin):
+class SearchIndexClient(HeadersMixin):
     """A client to interact with Azure search service Indexes.
 
     This class is not normally instantiated directly, instead use
@@ -37,7 +42,7 @@ class SearchIndexesClient(HeadersMixin):
     def __init__(self, endpoint, credential, **kwargs):
         # type: (str, AzureKeyCredential, **Any) -> None
 
-        self._endpoint = endpoint  # type: str
+        self._endpoint = normalize_endpoint(endpoint)  # type: str
         self._credential = credential  # type: AzureKeyCredential
         self._client = _SearchServiceClient(
             endpoint=endpoint, sdk_moniker=SDK_MONIKER, **kwargs
@@ -54,10 +59,20 @@ class SearchIndexesClient(HeadersMixin):
 
     async def close(self):
         # type: () -> None
-        """Close the :class:`~azure.search.documents.SearchIndexesClient` session.
+        """Close the :class:`~azure.search.documents.aio.SearchIndexClient` session.
 
         """
         return await self._client.close()
+
+    def get_search_client(self, index_name, **kwargs):
+        # type: (str, dict) -> SearchClient
+        """Return a client to perform operations on Search.
+
+        :param index_name: The name of the Search Index
+        :type index_name: str
+        :rtype: ~azure.search.documents.aio.SearchClient
+        """
+        return SearchClient(self._endpoint, index_name, self._credential, **kwargs)
 
     @distributed_trace_async
     async def list_indexes(self, **kwargs):
@@ -257,3 +272,165 @@ class SearchIndexesClient(HeadersMixin):
             index_name=index_name, request=analyze_request, **kwargs
         )
         return result
+
+    @distributed_trace_async
+    async def get_synonym_maps(self, **kwargs):
+        # type: (**Any) -> List[SynonymMap]
+        """List the Synonym Maps in an Azure Search service.
+
+        :return: List of synonym maps
+        :rtype: list[~azure.search.documents.indexes.models.SynonymMap]
+        :raises: ~azure.core.exceptions.HttpResponseError
+
+        .. admonition:: Example:
+
+            .. literalinclude:: ../samples/async_samples/sample_synonym_map_operations_async.py
+                :start-after: [START get_synonym_maps_async]
+                :end-before: [END get_synonym_maps_async]
+                :language: python
+                :dedent: 4
+                :caption: List Synonym Maps
+
+        """
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
+        result = await self._client.synonym_maps.list(**kwargs)
+        return [listize_synonyms(x) for x in result.synonym_maps]
+
+    @distributed_trace_async
+    async def get_synonym_map(self, name, **kwargs):
+        # type: (str, **Any) -> SynonymMap
+        """Retrieve a named Synonym Map in an Azure Search service
+
+        :param name: The name of the Synonym Map to get
+        :type name: str
+        :return: The retrieved Synonym Map
+        :rtype: ~azure.search.documents.indexes.models.SynonymMap
+        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        .. admonition:: Example:
+
+            .. literalinclude:: ../samples/async_samples/sample_synonym_map_operations_async.py
+                :start-after: [START get_synonym_map_async]
+                :end-before: [END get_synonym_map_async]
+                :language: python
+                :dedent: 4
+                :caption: Get a Synonym Map
+
+        """
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
+        result = await self._client.synonym_maps.get(name, **kwargs)
+        return listize_synonyms(result)
+
+    @distributed_trace_async
+    async def delete_synonym_map(self, synonym_map, **kwargs):
+        # type: (Union[str, SynonymMap], **Any) -> None
+        """Delete a named Synonym Map in an Azure Search service. To use access conditions,
+        the SynonymMap model must be provided instead of the name. It is enough to provide
+        the name of the synonym map to delete unconditionally.
+
+        :param name: The Synonym Map to delete
+        :type name: str or ~search.models.SynonymMap
+        :keyword match_condition: The match condition to use upon the etag
+        :type match_condition: ~azure.core.MatchConditions
+        :return: None
+        :rtype: None
+
+
+        .. admonition:: Example:
+
+            .. literalinclude:: ../samples/async_samples/sample_synonym_map_operations_async.py
+                :start-after: [START delete_synonym_map_async]
+                :end-before: [END delete_synonym_map_async]
+                :language: python
+                :dedent: 4
+                :caption: Delete a Synonym Map
+
+        """
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
+        error_map, access_condition = get_access_conditions(
+            synonym_map, kwargs.pop("match_condition", MatchConditions.Unconditionally)
+        )
+        kwargs.update(access_condition)
+        try:
+            name = synonym_map.name
+        except AttributeError:
+            name = synonym_map
+        await self._client.synonym_maps.delete(
+            synonym_map_name=name, error_map=error_map, **kwargs
+        )
+
+    @distributed_trace_async
+    async def create_synonym_map(self, name, synonyms, **kwargs):
+        # type: (str, Sequence[str], **Any) -> SynonymMap
+        """Create a new Synonym Map in an Azure Search service
+
+        :param name: The name of the Synonym Map to create
+        :type name: str
+        :param synonyms: A list of synonyms in SOLR format
+        :type synonyms: List[str]
+        :return: The created Synonym Map
+        :rtype: ~azure.search.documents.indexes.models.SynonymMap
+
+        .. admonition:: Example:
+
+            .. literalinclude:: ../samples/async_samples/sample_synonym_map_operations_async.py
+                :start-after: [START create_synonym_map_async]
+                :end-before: [END create_synonym_map_async]
+                :language: python
+                :dedent: 4
+                :caption: Create a Synonym Map
+
+        """
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
+        solr_format_synonyms = "\n".join(synonyms)
+        synonym_map = SynonymMap(name=name, synonyms=solr_format_synonyms)
+        result = await self._client.synonym_maps.create(synonym_map, **kwargs)
+        return listize_synonyms(result)
+
+    @distributed_trace_async
+    async def create_or_update_synonym_map(self, synonym_map, synonyms=None, **kwargs):
+        # type: (Union[str, SynonymMap], Optional[Sequence[str]], **Any) -> SynonymMap
+        """Create a new Synonym Map in an Azure Search service, or update an
+        existing one.
+
+        :param synonym_map: The name of the Synonym Map to create or update
+        :type synonym_map: str or ~azure.search.documents.SynonymMap
+        :param synonyms: A list of synonyms in SOLR format
+        :type synonyms: List[str]
+        :keyword match_condition: The match condition to use upon the etag
+        :type match_condition: ~azure.core.MatchConditions
+        :return: The created or updated Synonym Map
+        :rtype: ~azure.search.documents.indexes.models.SynonymMap
+
+        """
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
+        error_map, access_condition = get_access_conditions(
+            synonym_map, kwargs.pop("match_condition", MatchConditions.Unconditionally)
+        )
+        kwargs.update(access_condition)
+        try:
+            name = synonym_map.name
+            if synonyms:
+                synonym_map.synonyms = "\n".join(synonyms)
+            synonym_map.encryption_key = pack_search_resource_encryption_key(synonym_map.encryption_key)
+        except AttributeError:
+            name = synonym_map
+            solr_format_synonyms = "\n".join(synonyms)
+            synonym_map = SynonymMap(name=name, synonyms=solr_format_synonyms)
+        result = await self._client.synonym_maps.create_or_update(
+            synonym_map_name=name,
+            synonym_map=synonym_map,
+            error_map=error_map,
+            **kwargs
+        )
+        return listize_synonyms(result)
+
+    @distributed_trace_async
+    async def get_service_statistics(self, **kwargs):
+        # type: (**Any) -> dict
+        """Get service level statistics for a search service.
+
+        """
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
+        result = await self._client.get_service_statistics(**kwargs)
+        return result.as_dict()