+++ sdk/search/azure-search-documents/tests/async_tests/test_service_live_async.py
@@ -19,7 +19,7 @@ from search_service_preparer import SearchServicePreparer, SearchResourceGroupPr
 from azure_devtools.scenario_tests.utilities import trim_kwargs_from_test_function
 
 from azure.core.exceptions import HttpResponseError
-from azure.search.documents import(
+from azure.search.documents.indexes.models import(
     AnalyzeRequest,
     AnalyzeResult,
     CorsOptions,
@@ -27,19 +27,16 @@ from azure.search.documents import(
     SearchIndex,
     InputFieldMappingEntry,
     OutputFieldMappingEntry,
-    SearchServiceClient,
     ScoringProfile,
     SearchIndexerSkillset,
-    DataSourceCredentials,
-    SearchIndexerDataSource,
+    SearchIndexerDataSourceConnection,
     SearchIndexerDataContainer,
     SearchIndexer,
     SynonymMap,
     SimpleField,
     edm
 )
-from azure.search.documents.aio import SearchServiceClient
-from _test_utils import build_synonym_map_from_dict
+from azure.search.documents.indexes.aio import SearchIndexClient, SearchIndexerClient
 
 CWD = dirname(realpath(__file__))
 SCHEMA = open(join(CWD, "..", "hotel_schema.json")).read()
@@ -64,7 +61,7 @@ class SearchClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer()
     async def test_get_service_statistics(self, api_key, endpoint, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.get_service_statistics()
         assert isinstance(result, dict)
         assert set(result.keys()) == {"counters", "limits"}
@@ -74,7 +71,7 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer()
     async def test_list_indexes_empty(self, api_key, endpoint, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.list_indexes()
 
         with pytest.raises(StopAsyncIteration):
@@ -83,7 +80,7 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_list_indexes(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.list_indexes()
 
         first = await result.__anext__()
@@ -95,21 +92,21 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_index(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.get_index(index_name)
         assert result.name == index_name
 
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_index_statistics(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.get_index_statistics(index_name)
         assert set(result.keys()) == {'document_count', 'storage_size'}
 
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_indexes(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         await client.delete_index(index_name)
         import time
         if self.is_live:
@@ -121,7 +118,7 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_indexes_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
 
         # First create an index
         name = "hotels"
@@ -177,7 +174,7 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
             fields=fields,
             scoring_profiles=scoring_profiles,
             cors_options=cors_options)
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.create_index(index)
         assert result.name == "hotels"
         assert result.scoring_profiles[0].name == scoring_profile.name
@@ -200,7 +197,7 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
             fields=fields,
             scoring_profiles=scoring_profiles,
             cors_options=cors_options)
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.create_or_update_index(index_name=index.name, index=index)
         assert len(result.scoring_profiles) == 0
         assert result.cors_options.allowed_origins == cors_options.allowed_origins
@@ -223,7 +220,7 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_indexes_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
 
         # First create an index
         name = "hotels"
@@ -262,7 +259,7 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_analyze_text(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         analyze_request = AnalyzeRequest(text="One's <two/>", analyzer="standard.lucene")
         result = await client.analyze_text(index_name, analyze_request)
         assert len(result.tokens) == 2
@@ -271,14 +268,14 @@ class SearchSynonymMapsClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_synonym_map(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
         ])
-        assert isinstance(result, dict)
-        assert result["name"] == "test-syn-map"
-        assert result["synonyms"] == [
+        assert isinstance(result, SynonymMap)
+        assert result.name == "test-syn-map"
+        assert result.synonyms == [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
         ]
@@ -287,7 +284,7 @@ class SearchSynonymMapsClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_synonym_map(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
@@ -299,36 +296,35 @@ class SearchSynonymMapsClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_synonym_map_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         result = await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
         ])
-        sm_result = build_synonym_map_from_dict(result)
-        etag = sm_result.e_tag
+        etag = result.e_tag
 
         await client.create_or_update_synonym_map("test-syn-map", [
                     "Washington, Wash. => WA",
                 ])
 
-        sm_result.e_tag = etag
+        result.e_tag = etag
         with pytest.raises(HttpResponseError):
-            await client.delete_synonym_map(sm_result, match_condition=MatchConditions.IfNotModified)
+            await client.delete_synonym_map(result, match_condition=MatchConditions.IfNotModified)
             assert len(client.get_synonym_maps()) == 1
 
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_synonym_map(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
         ])
         assert len(await client.get_synonym_maps()) == 1
         result = await client.get_synonym_map("test-syn-map")
-        assert isinstance(result, dict)
-        assert result["name"] == "test-syn-map"
-        assert result["synonyms"] == [
+        assert isinstance(result, SynonymMap)
+        assert result.name == "test-syn-map"
+        assert result.synonyms == [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
         ]
@@ -336,7 +332,7 @@ class SearchSynonymMapsClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_synonym_maps(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         await client.create_synonym_map("test-syn-map-1", [
             "USA, United States, United States of America",
         ])
@@ -345,13 +341,13 @@ class SearchSynonymMapsClientTest(AzureMgmtTestCase):
         ])
         result = await client.get_synonym_maps()
         assert isinstance(result, list)
-        assert all(isinstance(x, dict) for x in result)
-        assert set(x['name'] for x in result) == {"test-syn-map-1", "test-syn-map-2"}
+        assert all(isinstance(x, SynonymMap) for x in result)
+        assert set(x.name for x in result) == {"test-syn-map-1", "test-syn-map-2"}
 
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_synonym_map(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
+        client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
         ])
@@ -361,9 +357,9 @@ class SearchSynonymMapsClientTest(AzureMgmtTestCase):
         ])
         assert len(await client.get_synonym_maps()) == 1
         result = await client.get_synonym_map("test-syn-map")
-        assert isinstance(result, dict)
-        assert result["name"] == "test-syn-map"
-        assert result["synonyms"] == [
+        assert isinstance(result, SynonymMap)
+        assert result.name == "test-syn-map"
+        assert result.synonyms == [
             "Washington, Wash. => WA",
         ]
 
@@ -372,7 +368,7 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_skillset(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
 
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
@@ -390,7 +386,7 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_skillset(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -405,7 +401,7 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_skillset_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -421,7 +417,7 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_skillset(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -439,7 +435,7 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_skillsets(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -453,7 +449,7 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_skillset(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -469,7 +465,7 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_skillset_inplace(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -485,7 +481,7 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_skillset_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -503,12 +499,11 @@ class SearchSkillsetClientTest(AzureMgmtTestCase):
 class SearchDataSourcesClientTest(AzureMgmtTestCase):
 
     def _create_datasource(self, name="sample-datasource"):
-        credentials = DataSourceCredentials(connection_string=CONNECTION_STRING)
         container = SearchIndexerDataContainer(name='searchcontainer')
-        data_source = SearchIndexerDataSource(
+        data_source = SearchIndexerDataSourceConnection(
             name=name,
             type="azureblob",
-            credentials=credentials,
+            connection_string=CONNECTION_STRING,
             container=container
         )
         return data_source
@@ -516,7 +511,7 @@ class SearchDataSourcesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_datasource_async(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         data_source = self._create_datasource()
         result = await client.create_datasource(data_source)
         assert result.name == "sample-datasource"
@@ -525,7 +520,7 @@ class SearchDataSourcesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_datasource_async(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         data_source = self._create_datasource()
         result = await client.create_datasource(data_source)
         assert len(await client.get_datasources()) == 1
@@ -535,7 +530,7 @@ class SearchDataSourcesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_datasource_async(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         data_source = self._create_datasource()
         created = await client.create_datasource(data_source)
         result = await client.get_datasource("sample-datasource")
@@ -544,7 +539,7 @@ class SearchDataSourcesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_list_datasource_async(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         data_source1 = self._create_datasource()
         data_source2 = self._create_datasource(name="another-sample")
         created1 = await client.create_datasource(data_source1)
@@ -556,7 +551,7 @@ class SearchDataSourcesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_datasource_async(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         data_source = self._create_datasource()
         created = await client.create_datasource(data_source)
         assert len(await client.get_datasources()) == 1
@@ -570,7 +565,7 @@ class SearchDataSourcesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_datasource_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         data_source = self._create_datasource()
         created = await client.create_datasource(data_source)
         etag = created.e_tag
@@ -589,7 +584,7 @@ class SearchDataSourcesClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_datasource_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         data_source = self._create_datasource()
         created = await client.create_datasource(data_source)
         etag = created.e_tag
@@ -609,16 +604,14 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     async def _prepare_indexer(self, endpoint, api_key, name="sample-indexer", ds_name="sample-datasource", id_name="hotels"):
         con_str = self.settings.AZURE_STORAGE_CONNECTION_STRING
         self.scrubber.register_name_pair(con_str, 'connection_string')
-        credentials = DataSourceCredentials(connection_string=con_str)
         container = SearchIndexerDataContainer(name='searchcontainer')
-        data_source = SearchIndexerDataSource(
+        data_source = SearchIndexerDataSourceConnection(
             name=ds_name,
             type="azureblob",
-            credentials=credentials,
+            connection_string=con_str,
             container=container
         )
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
-        ds_client = client.get_datasources_client()
+        ds_client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         ds = await ds_client.create_datasource(data_source)
 
         index_name = id_name
@@ -630,14 +623,14 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
           "searchable": False
         }]
         index = SearchIndex(name=index_name, fields=fields)
-        ind_client = client.get_indexes_client()
+        ind_client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
         ind = await ind_client.create_index(index)
         return SearchIndexer(name=name, data_source_name=ds.name, target_index_name=ind.name)
 
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_indexer(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         result = await client.create_indexer(indexer)
         assert result.name == "sample-indexer"
@@ -647,7 +640,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_indexer(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         result = await client.create_indexer(indexer)
         assert len(await client.get_indexers()) == 1
@@ -657,7 +650,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_indexer(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         created = await client.create_indexer(indexer)
         result = await client.get_indexer("sample-indexer")
@@ -666,7 +659,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_list_indexer(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer1 = await self._prepare_indexer(endpoint, api_key)
         indexer2 = await self._prepare_indexer(endpoint, api_key, name="another-indexer", ds_name="another-datasource", id_name="another-index")
         created1 = await client.create_indexer(indexer1)
@@ -678,7 +671,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_indexer(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         created = await client.create_indexer(indexer)
         assert len(await client.get_indexers()) == 1
@@ -692,7 +685,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_reset_indexer(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         result = await client.create_indexer(indexer)
         assert len(await client.get_indexers()) == 1
@@ -702,7 +695,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_run_indexer(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         result = await client.create_indexer(indexer)
         assert len(await client.get_indexers()) == 1
@@ -713,7 +706,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_indexer_status(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         result = await client.create_indexer(indexer)
         status = await client.get_indexer_status("sample-indexer")
@@ -722,7 +715,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_indexer_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         created = await client.create_indexer(indexer)
         etag = created.e_tag
@@ -738,7 +731,7 @@ class SearchIndexersClientTest(AzureMgmtTestCase):
     @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_indexer_if_unchanged(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexers_client()
+        client = SearchIndexerClient(endpoint, AzureKeyCredential(api_key))
         indexer = await self._prepare_indexer(endpoint, api_key)
         result = await client.create_indexer(indexer)
         etag = result.e_tag